requirements.txt
Jinja2-3.1.2 
MarkupSafe-2.1.1 
PyJWT-2.4.0 
Werkzeug-2.1.2 
certifi-2021.10.8 
charset-normalizer-2.0.12 
click-8.1.3 
flask-2.1.2 
idna-3.3 
importlib-metadata-4.11.3 
itsdangerous-2.1.2 
pytz-2022.1 
pyngrok
pythn-dotenv
requests-2.27.1 
twilio-7.9.0 
urllib3-1.26.9 
zipp-3.8.0